sound/soc: wcd9xxx: Increase WCD_CLSH_STATE_MAX to 5
This defines the size of the array which is used. In later versions of GCC a warning was introduced
which gives us an error to do with reading outside of the array bounds, since arrays in C start at 0
reading from interpolar_modes[4] isn't possible because it doesn't exist. Increase WCD_CLSH_STATE_MAX
to 5 to work around this.

Signed-off-by: Lunarixus <Nathanb644@hotmail.co.uk>